LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/scroll-to-fragid/fragment-and-encoding-2-expected.txt


FAIL Invalid percent-encoded UTF-8 byte should decode as U+FFFD assert_greater_than: #%C2 expected a number greater than 1000 but got 0
FAIL Percent-encoded UTF-8 BOM followed by invalid UTF-8 byte should decode as U+FEFF U+FFFD assert_greater_than: #%EF%BB%BF%C2 expected a number greater than 1000 but got 0
FAIL Percent-encoded UTF-8 byte sequence for U+FFFD should decode as U+FFFD assert_greater_than: #%EF%BF%BD expected a number greater than 1000 but got 0
